requirements.txt
# Core Flask and Web Framework
Flask==3.0.0

# OCR and Image Processing
pytesseract==0.3.10
Pillow>=10.0.0

# Document Processing
python-docx==1.1.0

# Excel Export
openpyxl==3.1.2

# HTTP Requests
requests>=2.31.0